refactor(control-planes): set-up data transformations
Set-up data transformations for the control-planes module.

Moves the data source for `/config` into the control-planes module as it belongs to it.
