+++ src/app/diagnostics/index.ts
@@ -1,5 +1,4 @@
 import { routes } from './routes'
-import { sources } from './sources'
 import type { ServiceDefinition } from '@/services/utils'
 import { token } from '@/services/utils'
 
@@ -7,15 +6,6 @@ type Token = ReturnType<typeof token>
 
 export const services = (app: Record<string, Token>): ServiceDefinition[] => {
   return [
-    [token('diagnostics.sources'), {
-      service: sources,
-      arguments: [
-        app.api,
-      ],
-      labels: [
-        app.sources,
-      ],
-    }],
     [token('diagnostics.routes'), {
       service: routes,
       labels: [
